% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce.R
\name{announce_post}
\alias{announce_post}
\title{Announce a Blog Post Across Multiple Platforms}
\usage{
announce_post(post)
}
\arguments{
\item{post}{A character vector of length
one, representing the file path to
the blog post. The function only supports
a single file path and uses the
first value if provided with more than
one.}
}
\value{
The function does not explicitly
return a value (\code{NULL}). It performs
side effects by creating posts on
external platforms and sending out a
newsletter.
}
\description{
The \code{announce_post} function automates the
task of announcing a blog post.
It extracts metadata from the post's YAML
front matter, generates relevant
messages, and posts announcements to
various platforms: Bluesky, LinkedIn,
Mastodon, and via newsletter.
}
\details{
\enumerate{
\item Checks that the \code{post} argument is
provided and raises errors or warnings
as needed.
\item Reads the YAML front matter of the blog
post to extract metadata such as
tags, slug, image, and additional fields
used for announcements.
\item Converts tags into a hashtag-based format.
\item Constructs a shortened URL for the blog
post based on its slug.
\item Posts the announcement to:
\itemize{
\item \strong{Bluesky}: Posts a truncated message
along with an image.
\item \strong{LinkedIn}: Posts the full summary
along with an image.
\item \strong{Mastodon}: Posts the SEO description,
image, and alternative text.
}
\item Sends the metadata to generate a
newsletter notification.
}
}
\examples{
\dontrun{
post <- "blog/post/2023-10-05-my-blog-post/index.md"
announce_post(post)
}
}
