% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{li_get_version}
\alias{li_get_version}
\title{Get the LinkedIn Version Date in YYYYMM Format}
\usage{
li_get_version()
}
\value{
A character string representing the
version date in \code{YYYYMM} format,
where \code{YYYY} is the 4-digit year, and \code{MM}
is the 2-digit month.
}
\description{
The \code{li_get_version} function calculates the
most recent past month's LinkedIn
version date and formats it as a string in
\code{YYYYMM} format. This date is used
to reference LinkedIn-specific API versions
or other functionalities tied to
monthly updates.
}
\details{
The function calculates the version date by
taking the current date and
rolling back to the previous month. The
result is then formatted as a string
in \code{YYYYMM} format using the \code{lubridate} and
\code{stringr} packages.
}
\examples{
\dontrun{
li_get_version()

# Store the version date in a variable
li_get_version()
}

}
