% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkedin.R
\name{li_urn_me}
\alias{li_urn_me}
\title{Retrieve the LinkedIn URN of the User}
\usage{
li_urn_me()
}
\value{
A string representing the user's
LinkedIn URN in the format \code{"urn:li:person:{id}"}.
}
\description{
This function fetches the LinkedIn Uniform
Resource Name (URN) for the currently
authenticated user using the LinkedIn API.
Authentication is performed using a
bearer token stored in the \code{LI_TOKEN}
environment variable.
}
\details{
The function sends an API request to the
LinkedIn \code{userinfo} endpoint with the
appropriate authorization token and extracts
the user's unique identifier (sub).
This identifier is then formatted as a
LinkedIn URN.
}
\examples{
\dontrun{
li_urn_me()
}
}
