% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_emoji}
\alias{random_emoji}
\title{Generate a Random Emoji}
\usage{
random_emoji()
}
\value{
A single character string
containing a randomly selected emoji.
}
\description{
The \code{random_emoji} function selects and
returns a random emoji from a predefined
list of emojis. This function is often
used to add personality or fun to
messages and announcements.
}
\details{
The function draws from a fixed list
of commonly used emojis. Each time the
function is called, it selects one emoji
at random and returns it. No input
parameters are required.
}
\examples{
# Generate a random emoji
random_emoji()

# Assign a random emoji to a variable
my_emoji <- random_emoji()
my_emoji

}
