% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{rladies_cols}
\alias{rladies_cols}
\title{RLadies+ brand colours}
\usage{
rladies_cols(...)
}
\arguments{
\item{...}{Character names of colours to retrieve.
If empty, returns all colours as a named character vector.}
}
\value{
A named character vector of hex colour values.
}
\description{
Access individual colours or the full palette from the RLadies+ brand.
}
\examples{
rladies_cols()
rladies_cols("purple", "rose", "blue")
rladies_cols("charcoal")
}
