% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{rladies_formats}
\alias{rladies_formats}
\alias{rladies_html}
\alias{rladies_pdf}
\title{RLadies+ R Markdown output formats}
\usage{
rladies_html(toc = TRUE, toc_float = TRUE, ...)

rladies_pdf(toc = FALSE, ...)
}
\arguments{
\item{toc}{Logical. Include a table of contents?}

\item{toc_float}{Logical. Float the table of contents (HTML only)?}

\item{...}{Additional arguments passed to the underlying format function.}
}
\description{
Output format functions for creating branded RLadies+ documents.
Use these in the \code{output} field of your R Markdown YAML header,
or select them from RStudio's "From Template" menu.
}
\examples{
\dontrun{
# In YAML header:
# output: spellbind::rladies_html
# output: spellbind::rladies_pdf
}
}
