% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{rladies_pal}
\alias{rladies_pal}
\title{RLadies+ colour palettes}
\usage{
rladies_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of the palette. One of \code{"main"}, \code{"full"}, \code{"purple"},
\code{"blue"}, \code{"rose"}, \code{"neutral"}, \code{"diverging"}, or \code{"light"}.}

\item{reverse}{Logical. Reverse the palette order?}

\item{...}{Additional arguments passed to \code{\link[grDevices:colorRamp]{grDevices::colorRampPalette()}}.}
}
\value{
A function that takes an integer \code{n} and returns \code{n} interpolated
colours.
}
\description{
Returns a colour palette function or vector from the RLadies+ brand.
}
\examples{
rladies_pal("main")(3)
rladies_pal("purple")(6)
rladies_pal("diverging", reverse = TRUE)(7)
}
