% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_rladies}
\alias{scale_rladies}
\alias{scale_colour_rladies}
\alias{scale_color_rladies}
\alias{scale_fill_rladies}
\alias{scale_colour_rladies_c}
\alias{scale_color_rladies_c}
\alias{scale_fill_rladies_c}
\title{RLadies+ colour and fill scales for ggplot2}
\usage{
scale_colour_rladies(palette = "main", reverse = FALSE, ...)

scale_color_rladies(palette = "main", reverse = FALSE, ...)

scale_fill_rladies(palette = "main", reverse = FALSE, ...)

scale_colour_rladies_c(palette = "purple", reverse = FALSE, ...)

scale_color_rladies_c(palette = "purple", reverse = FALSE, ...)

scale_fill_rladies_c(palette = "purple", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of the palette (see \code{\link[=rladies_pal]{rladies_pal()}} for options).}

\item{reverse}{Logical. Reverse the palette order?}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_colour_gradientn()}}.}
}
\description{
Discrete and continuous colour/fill scales using the RLadies+ brand
palettes.
}
\examples{
library(ggplot2)

ggplot(iris, aes(Sepal.Length, Sepal.Width, colour = Species)) +
  geom_point(size = 3) +
  scale_colour_rladies() +
  theme_rladies()

ggplot(faithfuld, aes(waiting, eruptions, fill = density)) +
  geom_tile() +
  scale_fill_rladies_c("purple") +
  theme_rladies()
}
