% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kit_newsletter.R
\name{send_newsletter}
\alias{send_newsletter}
\title{Send a Newsletter Using ConvertKit API}
\usage{
send_newsletter(frontmatter, url)
}
\arguments{
\item{frontmatter}{A named list containing
metadata extracted from the blog
post YAML. The list should include the
following fields:
\describe{
\item{\code{seo}}{A string containing the SEO
description for the blog post.}
\item{\code{image}}{A string with the relative
path to the blog post's thumbnail image.}
\item{\code{title}}{A string containing the
blog post's title.}
}}

\item{url}{Character, url of the post.}
}
\value{
This function does not explicitly
return a value (\code{NULL}). It performs
a side effect by sending a newsletter
through the ConvertKit API.
}
\description{
The \code{send_newsletter} function sends a
newsletter created using metadata
extracted from a blog post's front matter.
It communicates with the
ConvertKit API to create and schedule an
email broadcast.
}
\details{
The function generates the HTML content for
the newsletter using the \code{glue}
package (placeholders for custom templates
need to be defined). It then
constructs an HTTP POST request to the
ConvertKit API, passing various fields
such as the email \code{subject}, \code{thumbnail_url},
and \code{content}. The API key for
authentication is fetched from the
environment variable \code{KIT_SECRET}.
}
\examples{
\dontrun{
# Example frontmatter
frontmatter <- list(
  seo = "This is an amazing blog post about R programming!",
  image = "images/blog-thumbnail.jpg",
  title = "Learning R Programming"
)

# Sending a newsletter
send_newsletter(frontmatter)
}

}
