% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{short_url}
\alias{short_url}
\title{Shorten a URL using the Short.io API}
\usage{
short_url(uri)
}
\arguments{
\item{uri}{A character string. The
original URL to be shortened.}
}
\value{
A character string. The
shortened URL returned by the Short.io API.
}
\description{
\code{short_url} sends a POST request to the
Short.io API to create a new shortened
URL for the provided \code{uri}. The function
uses environment variable \code{SHORTIO}
for the API key and sets specific options
for the API call, such as not
allowing duplicate links and setting a
fixed domain for the shortened URL.
}
\details{
This function takes a given URL and
leverages the Short.io API to generate
a shortened version of the URL using
the specified custom domain.
}
\examples{
\dontrun{
  # Example usage to shorten a URL
  Sys.setenv(SHORTIO = "your_api_key_here") # Set API key in environment
  short_url("https://www.example.com/very-long-url")
}

}
